Make Option Permissive optional at build time
The way this is implemented changes the behavior on systems
with multiple cards, where one is supported and another is unsupported.

This sadly cannot be changed, because pciProbe is called before
preInit, so I can't know whether or not Option Permissive
is enabled while in pciProbe.

This commit adds a compile-time switch for whether or not
the code for Option Permissive is compiled, keeping old
behavior if desired, which is also the default for this option.

Signed-off-by: stefan11111 <stefan11111@shitposting.expert>